requirements.txt
torchsummary
scipy
tldextract
opencv-python
selenium
helium
selenium-wire
webdriver-manager
pandas
numpy
tqdm
Pillow
fvcore
pycocotools
scikit-learn
advertorch
python-telegram-bot
gspread
oauth2client




